refactor(app): restructure project to avoid circular imports
