+++ app/factory.py
@@ -1,0 +1,31 @@
+from fastapi import FastAPI
+from fastapi.middleware.cors import CORSMiddleware
+
+from app.core.config import settings
+from app.startup.lifecycle import lifespan
+
+
+def create_app() -> FastAPI:
+    """
+    创建并配置 FastAPI 应用实例。
+    """
+    app = FastAPI(
+        title=settings.PROJECT_NAME,
+        openapi_url=f"{settings.API_V1_STR}/openapi.json",
+        lifespan=lifespan
+    )
+
+    # 配置 CORS 中间件
+    app.add_middleware(
+        CORSMiddleware,
+        allow_origins=settings.ALLOWED_HOSTS,
+        allow_credentials=True,
+        allow_methods=["*"],
+        allow_headers=["*"],
+    )
+
+    return app
+
+
+# 创建 FastAPI 应用实例
+app = create_app()
